refactor: add unit tests for create project updates. split tests in multiple files for maintainability.
